WIP - Page Module Update
It's still under construction, but the goal is to support syntax highlights for the javascript code that will be used on the page.

In addition, possible support for JSDocs, to facilitate understanding the use of the functions offered by the App, such as: EmitEvent and Page (coming from the component of the same name).