Update of the Pages Module complete
Now, the pages are previously compiled and saved in memory, to optimize the server's response time.

The Pages API has been modified, and now, to declare the CSS of a page, you must create in the /pages directory, a .css file with the same name as the .html file to which you want to link.

For javascript, in addition to the well-known possibility of executing a server-side function before sending the response to the client, using `export function load()`, which is normally used to get data and send it along with the page's html, now we have the `export function mount()`.

In other words, just like the css that was extracted from the html, the javascript that must be executed on the page must be written within this mount function.