Developer Quality of Life Update
Now, 'ghost' functions and objects, such as EmitEvent and Pages, are defined in the Functions.js Module.

These can be normally imported within the pages' "mount()" scripts, to better control their use (with the help of the IDE).

The Develop.js project was also started, which is a module that aims to automatically update certain types, such as 'Pages' and 'Events', which are used in functions such as EmitEvent and emitRenderPageEvent (as agurment), thus making it clear to the developer his choices.