New Event System
The entire dynamic loading system has been rewritten,
for better security and performance.

Support for Events was also added in the Develop.js module, thus,
while the server is running, whenever a new
event is created, the module will update the list of available events,
used by the EmitEvent function. Therefore, we will always have help from IDE in
correct choice of which events there are to be executed.